cretae more pages and move pages into src
